+++ testutil/ante.go
@@ -2,6 +2,7 @@ package testutil
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/tabilabs/tabi/app"
 	"github.com/tabilabs/tabi/encoding"
 )
